# frontend/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "Evaluation": {
        "description": "A submitted evaluation as others see it (drafts are never shown to anyone else).",
        "properties": {
          "comment": {
            "title": "Comment",
            "type": "string"
          },
          "edited_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Last change after the first submission; show \"edited after submission\".",
            "title": "Edited At"
          },
          "evaluator": {
            "$ref": "#/components/schemas/UserRef"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "include_in_aggregate": {
            "description": "False for AI evaluations unless the owner includes them (Phase 6).",
            "title": "Include In Aggregate",
            "type": "boolean"
          },
          "is_ai": {
            "description": "The evaluator is an AI agent's service account (Phase 6).",
            "title": "Is Ai",
            "type": "boolean"
          },
          "recommendation": {
            "$ref": "#/components/schemas/Recommendation"
          },
          "scores": {
            "description": "Active criteria only, rubric order.",
            "items": {
              "$ref": "#/components/schemas/EvaluationScore"
            },
            "title": "Scores",
            "type": "array"
          },
          "submitted_at": {
            "description": "First submission.",
            "format": "date-time",
            "title": "Submitted At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "evaluator",
          "recommendation",
          "comment",
          "scores",
          "submitted_at",
          "edited_at",
          "is_ai",
          "include_in_aggregate"
        ],
        "title": "Evaluation",
        "type": "object"
      },
...
      "EvaluationList": {
        "description": "Submitted evaluations of an idea, or nothing while you owe your own.",
        "properties": {
          "items": {
            "description": "Oldest submission first; empty while score_hidden.",
            "items": {
              "$ref": "#/components/schemas/Evaluation"
            },
            "title": "Items",
            "type": "array"
          },
          "score_hidden": {
            "description": "You are an assigned evaluator who has not submitted (blind evaluation).",
            "title": "Score Hidden",
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "score_hidden"
        ],
        "title": "EvaluationList",
        "type": "object"
      },
...
      "EvaluationScore": {
        "properties": {
          "comment": {
            "title": "Comment",
            "type": "string"
          },
          "criterion_id": {
            "format": "uuid",
            "title": "Criterion Id",
            "type": "string"
          },
          "score": {
            "description": "Rubric score, 1 (low) to 5 (high).",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Score",
            "type": "integer"
          }
        },
        "required": [
          "criterion_id",
          "score",
          "comment"
        ],
        "title": "EvaluationScore",
        "type": "object"
      },
...
      "EvaluatorState": {
        "description": "An assigned evaluator's progress: no evaluation yet, draft saved, or submitted.",
        "enum": [
          "invited",
          "draft",
          "submitted"
        ],
        "title": "EvaluatorState",
        "type": "string"
      },
...
      "FieldError": {
        "description": "One request-validation error.",
        "properties": {
          "loc": {
            "description": "Location, e.g. ['body', 'title'].",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Loc",
            "type": "array"
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "FieldError",
        "type": "object"
      },
...
      "MyEvaluation": {
        "description": "Your evaluation of an idea you were asked to evaluate (the evaluate sheet).",
        "properties": {
          "comment": {
            "title": "Comment",
            "type": "string"
          },
          "due_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          },
          "editable": {
            "description": "Evaluation is open, so you can still save changes.",
            "title": "Editable",
            "type": "boolean"
          },
          "idea_id": {
            "format": "uuid",
            "title": "Idea Id",
            "type": "string"
          },
          "recommendation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Recommendation"
              },
              {
                "type": "null"
              }
            ]
          },
          "scores": {
            "description": "Saved scores for active criteria, rubric order.",
            "items": {
              "$ref": "#/components/schemas/MyScore"
            },
            "title": "Scores",
            "type": "array"
          },
          "state": {
            "$ref": "#/components/schemas/EvaluatorState",
            "description": "invited = nothing saved yet."
          },
          "submitted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Last save; null until first saved.",
            "title": "Updated At"
          }
        },
        "required": [
          "idea_id",
          "state",
          "editable",
          "due_at",
          "recommendation",
          "comment",
          "scores",
          "submitted_at",
          "updated_at"
        ],
        "title": "MyEvaluation",
        "type": "object"
      },
      "MyEvaluationIn": {
        "additionalProperties": false,
        "description": "Save your evaluation (replaces what was saved before).\n\n``submit: false`` saves a draft (any subset of scores). ``submit: true`` submits:\nevery active criterion needs a score and ``recommendation`` is required, else\n422 ``evaluation_incomplete``. A submitted evaluation stays submitted; later\nsaves must also use ``submit: true`` (409 ``evaluation_already_submitted``).",
        "properties": {
          "comment": {
            "default": "",
            "maxLength": 4000,
            "title": "Comment",
            "type": "string"
          },
          "recommendation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Recommendation"
              },
              {
                "type": "null"
              }
            ]
          },
          "scores": {
            "items": {
              "$ref": "#/components/schemas/MyScoreIn"
            },
            "maxItems": 6,
            "title": "Scores",
            "type": "array"
          },
          "submit": {
            "default": false,
            "title": "Submit",
            "type": "boolean"
          }
        },
        "title": "MyEvaluationIn",
        "type": "object"
      },
      "MyScore": {
        "properties": {
          "comment": {
            "title": "Comment",
            "type": "string"
          },
          "criterion_id": {
            "format": "uuid",
            "title": "Criterion Id",
            "type": "string"
          },
          "score": {
            "anyOf": [
              {
                "description": "Rubric score, 1 (low) to 5 (high).",
                "maximum": 5.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null only in a draft.",
            "title": "Score"
          }
        },
        "required": [
          "criterion_id",
          "score",
          "comment"
        ],
        "title": "MyScore",
        "type": "object"
      },
      "MyScoreIn": {
        "additionalProperties": false,
        "properties": {
          "comment": {
            "default": "",
            "maxLength": 1000,
            "title": "Comment",
            "type": "string"
          },
          "criterion_id": {
            "format": "uuid",
            "title": "Criterion Id",
            "type": "string"
          },
          "score": {
            "anyOf": [
              {
                "description": "Rubric score, 1 (low) to 5 (high).",
                "maximum": 5.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          }
        },
        "required": [
          "criterion_id"
        ],
        "title": "MyScoreIn",
        "type": "object"
      },
...
      "Problem": {
        "description": "RFC 9457 problem details, returned as ``application/problem+json``.\n\nClients should branch on ``code`` (stable, snake_case), not on ``title`` or\n``detail`` (human-readable, may change). As RFC 9457 allows, ``detail``,\n``instance`` and ``request_id`` may be absent: the one exception to \"every\ndocumented response field is always present\".",
        "examples": [
          {
            "code": "not_found",
            "detail": "Idea not found.",
            "instance": "/api/v1/ideas/0b7c7d1e-7a55-4a4f-9b8b-0d7d3a9d1c11",
            "request_id": "7f9c2b4e1d6a4c0f8e3b5a2d9c1e0f7a",
            "status": 404,
            "title": "Not Found",
            "type": "urn:soundings:problem:not_found"
          }
        ],
        "properties": {
          "code": {
            "description": "Stable machine-readable error code (snake_case).",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Explanation of this occurrence.",
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Request path that failed.",
            "title": "Instance"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Correlates with server logs (X-Request-ID).",
            "title": "Request Id"
          },
          "status": {
            "description": "HTTP status code.",
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Short, human-readable summary of the problem type.",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "URI identifying the problem type.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "Recommendation": {
        "enum": [
          "go",
          "maybe",
          "no"
        ],
        "title": "Recommendation",
        "type": "string"
      },
...
      "UserRef": {
        "description": "A user shown next to content: avatar, name.\n\nCarries no email; only people pickers (``UserSearchResult``), member lists\n(``Member``) and your own ``CurrentUser`` add it, for signed-in colleagues.",
        "properties": {
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "initials": {
            "description": "One or two letters for the avatar fallback.",
            "readOnly": true,
            "title": "Initials",
            "type": "string"
          }
        },
        "required": [
          "id",
          "display_name",
          "avatar_url",
          "initials"
        ],
        "title": "UserRef",
        "type": "object"
      },
...
      "ValidationProblem": {
        "description": "422 problem with per-field errors (``code`` is ``validation_error``).",
        "examples": [
          {
            "code": "not_found",
            "detail": "Idea not found.",
            "instance": "/api/v1/ideas/0b7c7d1e-7a55-4a4f-9b8b-0d7d3a9d1c11",
            "request_id": "7f9c2b4e1d6a4c0f8e3b5a2d9c1e0f7a",
            "status": 404,
            "title": "Not Found",
            "type": "urn:soundings:problem:not_found"
          }
        ],
        "properties": {
          "code": {
            "description": "Stable machine-readable error code (snake_case).",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Explanation of this occurrence.",
            "title": "Detail"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Errors",
            "type": "array"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Request path that failed.",
            "title": "Instance"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Correlates with server logs (X-Request-ID).",
            "title": "Request Id"
          },
          "status": {
            "description": "HTTP status code.",
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Short, human-readable summary of the problem type.",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "URI identifying the problem type.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "title": "ValidationProblem",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "session": {
        "description": "Server-side session cookie set by sign-in (Phase 1: POST /api/v1/auth/dev/login). Unsafe methods must also send the X-CSRF-Token header equal to the soundings_csrf cookie.",
        "in": "cookie",
        "name": "soundings_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Soundings REST API.\n\n* Errors are [RFC 9457](https://www.rfc-editor.org/rfc/rfc9457) problem details\n  (`application/problem+json`); branch on the `code` field.\n* Lists use cursor pagination: pass `next_cursor` back as `?cursor=`.\n* JSON fields are snake_case; timestamps are ISO 8601 with a UTC offset.\n",
    "title": "Soundings API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ideas/{idea}/evaluations": {
      "get": {
        "description": "Blind: empty with score_hidden true while you are an assigned evaluator who has not submitted. Otherwise every submitted evaluation. Never drafts.",
        "operationId": "list_evaluations",
        "parameters": [
          {
            "description": "The idea: its id (UUID) or its key such as \"CUST-12\" (any case).",
            "in": "path",
            "name": "idea",
            "required": true,
            "schema": {
              "description": "The idea: its id (UUID) or its key such as \"CUST-12\" (any case).",
              "maxLength": 36,
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}|[A-Za-z][A-Za-z0-9]{1,5}-[1-9][0-9]{0,8})$",
              "title": "Idea",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationList"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not signed in (unauthorized)"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found, or not visible to you (not_found)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "description": "Validation Failed"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "List submitted evaluations",
        "tags": [
          "evaluations"
        ]
      }
    },
    "/api/v1/ideas/{idea}/evaluations/me": {
      "get": {
        "description": "The evaluate sheet's state; null if you are not an evaluator of this idea.",
        "operationId": "get_my_evaluation",
        "parameters": [
          {
            "description": "The idea: its id (UUID) or its key such as \"CUST-12\" (any case).",
            "in": "path",
            "name": "idea",
            "required": true,
            "schema": {
              "description": "The idea: its id (UUID) or its key such as \"CUST-12\" (any case).",
              "maxLength": 36,
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}|[A-Za-z][A-Za-z0-9]{1,5}-[1-9][0-9]{0,8})$",
              "title": "Idea",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/MyEvaluation"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get My Evaluation"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not signed in (unauthorized)"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found, or not visible to you (not_found)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "description": "Validation Failed"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Get my evaluation",
        "tags": [
          "evaluations"
        ]
      },
      "put": {
        "description": "Replaces your saved evaluation. 403 unless you are an assigned evaluator with role member/admin. 409 evaluation_closed unless evaluation is open. 409 evaluation_already_submitted for submit=false after submitting. 422 evaluation_incomplete (errors list what is missing). 422 unknown_criterion for a criterion outside the active rubric.",
        "operationId": "save_my_evaluation",
        "parameters": [
          {
            "description": "The idea: its id (UUID) or its key such as \"CUST-12\" (any case).",
            "in": "path",
            "name": "idea",
            "required": true,
            "schema": {
              "description": "The idea: its id (UUID) or its key such as \"CUST-12\" (any case).",
              "maxLength": 36,
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}|[A-Za-z][A-Za-z0-9]{1,5}-[1-9][0-9]{0,8})$",
              "title": "Idea",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MyEvaluationIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyEvaluation"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not signed in (unauthorized)"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not allowed (forbidden, csrf_failed, or a specific code)"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found, or not visible to you (not_found)"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflicts with the current state (see code)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "description": "Validation Failed"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Save or submit my evaluation",
        "tags": [
          "evaluations"
        ]
      }
    },
...
  }
}